Download ERA5 reanalysis of temperature for a given period and area data is stored per month in a netcdf file /!\ download quite long...

Download ERA5 reanalysis of temperature for a given period and area data is stored per month in a netcdf file /!\ download quite long...

download_era5(user_id, years, months, latn, latx, lonn, lonx, storage_dir)

Arguments

user_id

character with user Copernicus UID

years

vector of years to download

months

vector of characters for months to download (e.g. c("01", "02", "03")).

latn

minimum latitude

latx

maximum latitude

lonn

minimum longitude

lonx

maximum longitude

storage_dir

directory to save the downloaded files